app/requirements.txt
typing
sympy
re
python-dotenv
langchain-openai